backend/requirements.txt

Flask==2.3.2
Flask-CORS==3.0.10
mysql-connector-python==8.0.33
transformers==4.30.2
torch==2.0.1
scikit-learn==1.2.2
joblib==1.2.0
tensorflow==2.12.0
nltk==3.8.1
twilio==8.3.0